refactor(storyboard): #42 - 사이드바 레이아웃 구조 변경
