+++ app/(tabs)/storyboard/drawer/drawer-search.js
@@ -1,0 +1,13 @@
+import { StyleSheet, Text, View } from 'react-native';
+
+const DrawerSearch = () => {
+  return (
+    <View>
+      <Text>drawer-search</Text>
+    </View>
+  );
+};
+
+export default DrawerSearch;
+
+const styles = StyleSheet.create({});
